ccna-tutor/requirements.txt
flask==3.0.0
gunicorn==21.2.0
anthropic==0.18.1
python-dotenv==1.2.1
